# Optimize various string operations