… Tests/FluentAssertions.Specs/Formatting/MultidimensionalArrayFormatterSpecs.cs