Three little changes to move/remove some cruft when working with strings.

For the first commit I benchmarked the difference between several alternatives and the most primitive one was the fastest one.

<details>
<summary>Benchmark results</summary>

| Method           | MemberName       | Mean       | Error      | StdDev     | Median     | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|----------------- |----------------- |-----------:|-----------:|-----------:|-----------:|------:|--------:|-------:|-------:|----------:|------------:|
| RegexInstance    | Parent.Property2 | 820.262 ns | 15.9691 ns | 22.3865 ns | 816.300 ns | 1.001 |    0.04 | 0.3834 | 0.0019 |    2408 B |        1.00 |
| RegexStatic      | Parent.Property2 |  72.595 ns |  0.9518 ns |  0.7948 ns |  72.406 ns | 0.089 |    0.00 |      - |      - |         - |        0.00 |
| RegexCache       | Parent.Property2 |  64.338 ns |  1.0634 ns |  0.9947 ns |  64.712 ns | 0.078 |    0.00 |      - |      - |         - |        0.00 |
| RegexSource      | Parent.Property2 |  32.408 ns |  0.6633 ns |  0.7373 ns |  31.989 ns | 0.040 |    0.00 |      - |      - |         - |        0.00 |
| LinqAndPattern   | Parent.Property2 |  17.248 ns |  0.2948 ns |  0.2757 ns |  17.262 ns | 0.021 |    0.00 | 0.0051 |      - |      32 B |        0.01 |
| IndexOfAny       | Parent.Property2 |  10.781 ns |  0.0412 ns |  0.0344 ns |  10.770 ns | 0.013 |    0.00 | 0.0051 |      - |      32 B |        0.01 |
| IndexOfAnyCached | Parent.Property2 |   6.062 ns |  0.1416 ns |  0.2365 ns |   5.933 ns | 0.007 |    0.00 |      - |      - |         - |        0.00 |
| Contains         | Parent.Property2 |   1.966 ns |  0.0569 ns |  0.0475 ns |   1.974 ns | 0.002 |    0.00 |      - |      - |         - |        0.00 |
| ContainsOrdinal  | Parent.Property2 |   2.396 ns |  0.0143 ns |  0.0127 ns |   2.397 ns | 0.003 |    0.00 |      - |      - |         - |        0.00 |
|                  |                  |            |            |            |            |       |         |        |        |           |             |
| RegexInstance    | Property1        | 783.972 ns |  3.8892 ns |  3.2477 ns | 783.435 ns | 1.000 |    0.01 | 0.3748 | 0.0019 |    2352 B |        1.00 |
| RegexStatic      | Property1        |  43.992 ns |  0.1414 ns |  0.1253 ns |  44.002 ns | 0.056 |    0.00 |      - |      - |         - |        0.00 |
| RegexCache       | Property1        |  41.948 ns |  0.8558 ns |  1.5432 ns |  42.259 ns | 0.054 |    0.00 |      - |      - |         - |        0.00 |
| RegexSource      | Property1        |  30.557 ns |  0.6352 ns |  0.9110 ns |  30.647 ns | 0.039 |    0.00 |      - |      - |         - |        0.00 |
| LinqAndPattern   | Property1        |  21.860 ns |  0.1412 ns |  0.1252 ns |  21.860 ns | 0.028 |    0.00 | 0.0051 |      - |      32 B |        0.01 |
| IndexOfAny       | Property1        |  10.921 ns |  0.2096 ns |  0.1858 ns |  10.897 ns | 0.014 |    0.00 | 0.0051 |      - |      32 B |        0.01 |
| IndexOfAnyCached | Property1        |   5.801 ns |  0.1231 ns |  0.1152 ns |   5.738 ns | 0.007 |    0.00 |      - |      - |         - |        0.00 |
| Contains         | Property1        |   6.448 ns |  0.0354 ns |  0.0296 ns |   6.444 ns | 0.008 |    0.00 |      - |      - |         - |        0.00 |
| ContainsOrdinal  | Property1        |   6.694 ns |  0.1196 ns |  0.0999 ns |   6.656 ns | 0.009 |    0.00 |      - |      - |         - |        0.00 |

</details>

<details>
<summary>Benchmark code</summary>

```cs
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Text.RegularExpressions;

BenchmarkRunner.Run<Benchmark>();

[MemoryDiagnoser]
public class Benchmark
{
    [Params("Property1", "Parent.Property2")]
    public string MemberName { get; set; }

    [Benchmark(Baseline = true)]
    public bool RegexInstance() => new Regex(@"[\.\[\]]").IsMatch(MemberName);

    [Benchmark]
    public bool RegexStatic() => Regex.IsMatch(MemberName, @"[\.\[\]]");

    private static readonly Regex regex = new Regex(@"[\.\[\]]");

    [Benchmark]
    public bool RegexCache() => regex.IsMatch(MemberName);

    [Benchmark]
    public bool RegexSource() => A.Contains().IsMatch(MemberName);

    [Benchmark]
    public bool LinqAndPattern() => MemberName.Any(e => e is '.' or '[' or ']');

    [Benchmark]
    public bool IndexOfAny() => MemberName.IndexOfAny(['.', '[', ']']) >= 0;

    private static readonly char[] chars = ['.', '[', ']'];

    [Benchmark]
    public bool IndexOfAnyCached() => MemberName.IndexOfAny(chars) >= 0;

    [Benchmark]
    public bool Contains() => MemberName.Contains('.') || MemberName.Contains('[') || MemberName.Contains(']');

    [Benchmark]
    public bool ContainsOrdinal() =>
        MemberName.Contains('.', StringComparison.Ordinal) || MemberName.Contains('[', StringComparison.Ordinal) || MemberName.Contains(']', StringComparison.Ordinal);
}

static partial class A
{
    [GeneratedRegex(@"[\.\[\]]")]
    public static partial Regex Contains();
}
```

</details>

The second and third commit uses better overloads of `string.Replace` and `string.Split`

## IMPORTANT 

* [x] If the PR touches the public API, the changes have been approved in a separate issue with the "api-approved" label.
* [x] The code complies with the [Coding Guidelines for C#](https://www.csharpcodingguidelines.com/).
* [x] The changes are covered by unit tests which follow the Arrange-Act-Assert syntax and the naming conventions such as is used [in these tests](../tree/develop/Tests/FluentAssertions.Equivalency.Specs/MemberMatchingSpecs.cs#L51-L430).
* [x] If the PR adds a feature or fixes a bug, please update [the release notes](../tree/develop/docs/_pages/releases.md) with a functional description that explains what the change means to consumers of this library, which are published on the [website](https://fluentassertions.com/releases).
* [x] If the PR changes the public API the changes needs to be included by running [AcceptApiChanges.ps1](../tree/develop/AcceptApiChanges.ps1) or [AcceptApiChanges.sh](../tree/develop/AcceptApiChanges.sh).
* [x] If the PR affects [the documentation](../tree/develop/docs/_pages), please include your changes in this pull request so the documentation will appear on the [website](https://www.fluentassertions.com/introduction).
    * [x] Please also run `./build.sh --target spellcheck` or `.\build.ps1 --target spellcheck` before pushing and check the good outcome