@jnyrup 2025-03-02
> The title doesn't seem to cover the purpose of this PR, right? It's about performance optimizations and not about text cleanup.

Thanks, updated the title to clarify what's going on.